chat_integration/app/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform>=1.48.0
requests
